Move application attribute settings logic to a separate hook
