+++ features/admin.applications.v1/components/settings/attribute-management/attribute-settings.tsx
@@ -51,6 +51,7 @@ import { AttributeSelection } from "./attribute-selection";
 import { AttributeSelectionOIDC } from "./attribute-selection-oidc";
 import { RoleMapping } from "./role-mapping";
 import { updateAuthProtocolConfig, updateClaimConfiguration } from "../../../api/application";
+import useApplicationAttributeSettings from "../../../hooks/use-application-attribute-settings";
 import {
     AppClaimInterface,
     ClaimConfigurationInterface,
@@ -177,6 +178,15 @@ export const AttributeSettings: FunctionComponent<AttributeSettingsPropsInterfac
         [ "data-componentid" ]: componentId
     } = props;
 
+    const {
+        getExternalClaimsGroupedByScopes,
+        getMappedClaims,
+        isClaimLoading
+    } = useApplicationAttributeSettings(
+        claimConfigurations,
+        onlyOIDCConfigured
+    );
+
     const { t } = useTranslation();
 
     const dispatch: Dispatch = useDispatch();
@@ -189,8 +199,7 @@ export const AttributeSettings: FunctionComponent<AttributeSettingsPropsInterfac
 
     const [ selectedDialect, setSelectedDialect ] = useState<SelectedDialectInterface>();
 
-    // Get OIDC scope-use attributes list
-    const [ scopes, setScopes ] = useState<OIDCScopesListInterface[]>(null);
+
     const [ externalClaimsGroupedByScopes, setExternalClaimsGroupedByScopes ]
         = useState<OIDCScopesClaimsListInterface[]>(null);
     const [ unfilteredExternalClaimsGroupedByScopes, setUnfilteredExternalClaimsGroupedByScopes ]
@@ -199,8 +208,6 @@ export const AttributeSettings: FunctionComponent<AttributeSettingsPropsInterfac
     // Manage available claims in local and external dialects.
     const [ isClaimRequestLoading, setIsClaimRequestLoading ] = useState(true);
     const [ claims, setClaims ] = useState<ExtendedClaimInterface[]>([]);
-    const [ externalClaims, setExternalClaims ] = useState<ExtendedExternalClaimInterface[]>([]);
-    const [ unfilteredExternalClaims, setUnfilteredExternalClaims ] = useState<ExtendedExternalClaimInterface[]>([]);
     const [ isScopeExternalClaimMappingLoading, setIsScopeExternalClaimMappingLoading ] = useState(true);
 
     // Selected claims in local and external dialects.
@@ -221,57 +228,11 @@ export const AttributeSettings: FunctionComponent<AttributeSettingsPropsInterfac
     // Role Mapping.
     const [ roleMapping, setRoleMapping ] = useState<RoleMappingInterface[]>(claimConfigurations?.role?.mappings ?? []);
 
-    const [ isClaimLoading, setIsClaimLoading ] = useState<boolean>(true);
     const [ isUserAttributesLoading, setUserAttributesLoading ] = useState<boolean>(undefined);
 
     const eventPublisher: EventPublisher = EventPublisher.getInstance();
 
-    const {
-        data: OIDCScopeList,
-        isLoading: isOIDCScopeListLoading,
-        isValidating: isOIDCScopeListValidating
-    } = useOIDCScopesList();
-
     const [ duplicatedMappingValues,setDuplicatedMappingValues ] = useState<Array<string>>([]);
-
-    /**
-     * Get local mapped claim display name for external claims
-     */
-    useEffect(() => {
-        const filteredExternalClaims: ExtendedExternalClaimInterface[] = unfilteredExternalClaims
-            .filter((claim: ExtendedExternalClaimInterface) => {
-                const matchedLocalClaim: ExtendedClaimInterface[] = claims
-                    .filter((localClaim: ExtendedClaimInterface) => {
-                        return localClaim.claimURI === claim.mappedLocalClaimURI;
-                    });
-
-                return matchedLocalClaim.length !== 0;
-            });
-
-        filteredExternalClaims.forEach((externalClaim: ExtendedExternalClaimInterface) => {
-            const mappedLocalClaimUri: string = externalClaim.mappedLocalClaimURI;
-            const matchedLocalClaim: ExtendedClaimInterface[] = claims
-                .filter((localClaim: ExtendedClaimInterface) => {
-                    return localClaim.claimURI === mappedLocalClaimUri;
-                });
-
-            if (matchedLocalClaim && matchedLocalClaim[0] && matchedLocalClaim[0].displayName) {
-                externalClaim.localClaimDisplayName = matchedLocalClaim[0].displayName;
-            }
-        });
-        setExternalClaims(filteredExternalClaims);
-    }, [ claims, unfilteredExternalClaims ]);
-
-    useEffect(() => {
-        if (!isOIDCScopeListLoading && !isOIDCScopeListValidating) {
-            setScopes(OIDCScopeList);
-            getClaims();
-            getAllDialects();
-            findLocalClaimDialectURI();
-
-            return;
-        }
-    }, [ isOIDCScopeListLoading, isOIDCScopeListValidating, OIDCScopeList ]);
 
     useEffect(() => {
         getExternalClaimsGroupedByScopes();
@@ -317,167 +278,6 @@ export const AttributeSettings: FunctionComponent<AttributeSettingsPropsInterfac
     }, [ claimConfigurations ]);
 
     /**
-     * Check whether claim is mandatory or not
-     *
-     * @param uri - Claim URI to be checked.
-     *
-     * @returns If initially requested as mandatory.
-     */
-    const checkInitialRequestMandatory = (uri: string): boolean => {
-        const externalClaim: ExtendedExternalClaimInterface = externalClaims
-            .find((claim: ExtendedExternalClaimInterface) => claim.claimURI === uri);
-
-        if (externalClaim){
-            const requestURI: boolean = claimConfigurations.requestedClaims.find(
-                (requestClaims: RequestedClaimConfigurationInterface) => (
-                    requestClaims?.claim?.uri === externalClaim.mappedLocalClaimURI))?.mandatory;
-
-            if (requestURI !== undefined) {
-                return requestURI;
-            }
-        }
-
-        return false;
-    };
-
-    /**
-     * Check whether claim is requested or not.
-     *
-     * @param uri - Claim URI to be checked.
-     *
-     * @returns If initially requested or not.
-     */
-    const checkInitialRequested = (uri: string): boolean => {
-        const externalClaim: ExtendedExternalClaimInterface = externalClaims
-            .find((claim: ExtendedExternalClaimInterface) => claim.claimURI === uri);
-
-        if (externalClaim){
-            const requestURI: RequestedClaimConfigurationInterface = claimConfigurations.requestedClaims
-                .find((requestClaims: RequestedClaimConfigurationInterface) =>
-                    requestClaims?.claim?.uri === externalClaim.mappedLocalClaimURI);
-
-            return requestURI !== undefined;
-        }
-
-        return false;
-    };
-
-    /**
-     * Grouped Scopes and external claims
-     */
-    const getExternalClaimsGroupedByScopes = () => {
-        const tempClaims: ExtendedExternalClaimInterface[] = [ ...externalClaims ];
-        const updatedScopes: OIDCScopesClaimsListInterface[] = [];
-        const scopedClaims: ExtendedExternalClaimInterface[] = [];
-
-        if ((scopes !== null) && (scopes !== undefined) && (scopes.length !== 0) && (claims.length !== 0)) {
-            scopes.map((scope: OIDCScopesListInterface) => {
-                if (scope.name !== "openid"){
-                    const updatedClaims: ExtendedExternalClaimInterface[] = [];
-                    let scopeSelected: boolean = false;
-
-                    scope.claims.map((scopeClaim: string) => {
-                        tempClaims.map( (tempClaim: ExtendedExternalClaimInterface) => {
-                            if (scopeClaim === tempClaim.claimURI) {
-                                const updatedClaim: ExtendedExternalClaimInterface = {
-                                    ...tempClaim,
-                                    mandatory: checkInitialRequestMandatory(tempClaim.claimURI),
-                                    requested: checkInitialRequested(tempClaim.claimURI)
-                                };
-
-                                if (updatedClaim.requested) {
-                                    scopeSelected = true;
-                                }
-                                updatedClaims.push(updatedClaim);
-                                scopedClaims.push(tempClaim);
-                            }
-                        });
-                    });
-                    const updatedScope: OIDCScopesClaimsListInterface = {
-                        ...scope,
-                        claims: updatedClaims,
-                        selected: scopeSelected
-                    };
-
-                    updatedScopes.push(updatedScope);
-                }
-            });
-
-            const scopelessClaims: ExtendedExternalClaimInterface[] = tempClaims.filter(
-                (tempClaim: ExtendedExternalClaimInterface) => !scopedClaims.includes(tempClaim));
-            const updatedScopelessClaims: ExtendedExternalClaimInterface[] = [];
-            let isScopelessClaimRequested: boolean = false;
-
-            scopelessClaims.map((tempClaim: ExtendedExternalClaimInterface) => {
-                const isInitialRequested: boolean = checkInitialRequested(tempClaim.claimURI);
-
-                updatedScopelessClaims.push({
-                    ...tempClaim,
-                    mandatory: checkInitialRequestMandatory(tempClaim.claimURI),
-                    requested: checkInitialRequested(tempClaim.claimURI)
-                });
-                isScopelessClaimRequested = isInitialRequested;
-            });
-            updatedScopes.push({
-                claims: updatedScopelessClaims,
-                description: t("applications:edit.sections.attributes" +
-                    ".selection.scopelessAttributes.description"),
-                displayName: t("applications:edit.sections.attributes" +
-                    ".selection.scopelessAttributes.displayName"),
-                name: t("applications:edit.sections.attributes" +
-                    ".selection.scopelessAttributes.name"),
-                selected: isScopelessClaimRequested
-            });
-        }
-
-        // Derive scopes for markdown guide.
-        if (onlyOIDCConfigured) {
-            const derivedScopesList: string[] = [ "openid" ];
-
-            updatedScopes.forEach((scope: OIDCScopesClaimsListInterface) => {
-                if (scope.selected && scope.name !== "openid" && scope.name !== "scopeless") {
-                    derivedScopesList.push(scope.name);
-                }
-            });
-            dispatch(setApplicationScopes(derivedScopesList.join(" ")));
-        }
-
-        setUnfilteredExternalClaimsGroupedByScopes(sortBy(updatedScopes, "name"));
-        setExternalClaimsGroupedByScopes(sortBy(updatedScopes, "name"));
-        setIsScopeExternalClaimMappingLoading(false);
-    };
-
-    /**
-     * Get Local Claims
-     */
-    const getClaims = () => {
-        setIsClaimLoading(true);
-        const params: ClaimsGetParams = {
-            "exclude-hidden-claims": true,
-            filter: null,
-            limit: null,
-            offset: null,
-            sort: null
-        };
-
-        getAllLocalClaims(params)
-            .then((response: Claim[]) => {
-                setClaims(response);
-            })
-            .catch(() => {
-                dispatch(addAlert({
-                    description: t("claims:local.notifications.fetchLocalClaims.genericError" +
-                        ".description"),
-                    level: AlertLevels.ERROR,
-                    message: t("claims:local.notifications.fetchLocalClaims." +
-                        "genericError.message")
-                }));
-            }).finally(() => {
-                setIsClaimLoading(false);
-            });
-    };
-
-    /**
      * Get All Dialected
      */
     const getAllDialects = () => {
@@ -494,32 +294,6 @@ export const AttributeSettings: FunctionComponent<AttributeSettingsPropsInterfac
                         "genericError.message")
                 }));
             });
-    };
-
-    /**
-     * Get All External Claims
-     */
-    const getMappedClaims = (newClaimId: string) => {
-        if (newClaimId !== null) {
-            getAllExternalClaims(newClaimId, null)
-                .then((response: ExternalClaim[]) => {
-                    setIsClaimRequestLoading(true);
-                    setIsScopeExternalClaimMappingLoading(true);
-                    setUnfilteredExternalClaims(response);
-                })
-                .catch(() => {
-                    dispatch(addAlert({
-                        description: t("claims:external.notifications.fetchExternalClaims" +
-                            ".genericError.description"),
-                        level: AlertLevels.ERROR,
-                        message: t("claims:external.notifications.fetchExternalClaims" +
-                            ".genericError.message")
-                    }));
-                })
-                .finally(() => {
-                    setIsClaimRequestLoading(false);
-                });
-        }
     };
 
     const findDialectID = (value: string) => {
